Update DaryCard using bloc
