+++ lib/ui/diary_home_page/diary_home_event.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+@immutable
+abstract class DiaryHomeEvent {}
+
+class InputVisibleEvent extends DiaryHomeEvent {
+  InputVisibleEvent();
+}
+
