+++ lib/ui/diary_home_page/diary_home_provider.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import 'diary_home_bloc.dart';
+import 'diary_home_view.dart';
+
+class DiaryHomeProvider extends BlocProvider<DiaryHomeBloc> {
+
+  DiaryHomeProvider({
+    Key? key,
+  }) : super(
+          key: key,
+          create: (context) => DiaryHomeBloc(),
+          // child: const DiaryHomeScreen(),
+        );
+}
